PR Feedback: Fix tests
